app/requirements.txt
aio-pika==9.3.0
aiohttp==3.8.6
aiormq==6.7.7
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
dnspython==2.4.2
ecdsa==0.18.0
elastic-transport==8.10.0
elasticsearch==8.11.0
exceptiongroup==1.1.3
fastapi==0.104.1
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
importlib-resources==6.1.1
Jinja2==3.1.2
limits==3.6.0
loguru==0.7.2
MarkupSafe==2.1.3
motor==3.3.2
multidict==6.0.4
packaging==23.2
pamqp==3.2.1
passlib==1.7.4
pika==1.3.2
pyasn1==0.5.0
pydantic==2.5.1
pydantic_core==2.14.3
pymongo==4.6.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
rsa==4.9
six==1.16.0
slowapi==0.1.8
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
wrapt==1.16.0
yarl==1.9.2
